Renamed to *.php from *.php3 per recommendations for PHP4. Changes to support new database API for MySQL and PostgreSQL support. SQL change, reword left join as predicate-based join. Don't assume area 1 exists, and don't force area to 0. Better SQL error handling; always present full HTML pages on errors. Don't quote numeric values in SQL. Different but equivalent fix as 0.9pre2 for I18 time units and weekday names. Enhancements to support up to 10 site-configured meeting types. Enhancement to 0.0pre2 fix so week can start on any day. HTML fixups: quote tag attributes where needed by HTML standard.
Don't use stripslashes on database records; the extra backslashes this was
meant to remove should not have been in there in the first place.

Remove code which tried to intelligently report start and end date and
time based on duration.  Sometimes it omitted the end date when it should
have been there, for example a 4 hour meeting which crossed midnight.  Now,
always report both date and time for both start and end.